use the pdfjs that comes with obsidian
