+++ lib/package.json
@@ -50,7 +50,6 @@
     "@apollo/utils.createhash": "^3.0.0",
     "mammoth": "^1.6.0",
     "p-queue": "^7.4.1",
-    "pdfjs-dist": "^4.2.67",
     "pure-md5": "^0.1.14",
     "tesseract.js": "^3.0.3",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.1/xlsx-0.20.1.tgz"
