+++ lib/src/pdf/pdf-manager.ts
@@ -1,56 +1,11 @@
-import { Platform, TFile } from 'obsidian'
+import { Platform, TFile, loadPdfJs } from 'obsidian'
 import WebWorker from 'web-worker:./pdf-worker.ts'
 import {
   CANT_EXTRACT_ON_MOBILE,
   FAILED_TO_EXTRACT,
   pdfProcessQueue,
-  workerTimeout,
 } from '../globals'
 import { getCachePath, readCache, writeCache } from '../cache'
-
-class PDFWorker {
-  static #pool: PDFWorker[] = []
-  #running = false
-
-  private constructor(private worker: Worker) { }
-
-  static getWorker(): PDFWorker {
-    const free = PDFWorker.#pool.find(w => !w.#running)
-    if (free) {
-      return free
-    }
-    // Spawn a new worker
-    const worker = new PDFWorker(new WebWorker({ name: 'PDF Text Extractor' }))
-    PDFWorker.#pool.push(worker)
-    return worker
-  }
-
-  static #destroyWorker(pdfWorker: PDFWorker) {
-    pdfWorker.worker.terminate()
-    PDFWorker.#pool = PDFWorker.#pool.filter(w => w !== pdfWorker)
-  }
-
-  public async run(msg: { data: string; name: string }): Promise<any> {
-    console.log("Running PDFWorker for " + msg.name)
-    return new Promise((resolve, reject) => {
-      this.#running = true
-
-      const timeout = setTimeout(() => {
-        console.warn('Text Extractor - PDF Worker timeout for ', msg.name)
-        reject('timeout')
-        PDFWorker.#destroyWorker(this)
-      }, workerTimeout)
-
-      this.worker.postMessage(msg)
-      this.worker.onmessage = evt => {
-        console.log("PDFWorker finished for " + msg.name)
-        clearTimeout(timeout)
-        resolve(evt)
-        this.#running = false
-      }
-    })
-  }
-}
 
 class PDFManager {
   public async getPdfText(file: TFile): Promise<string> {
@@ -78,31 +33,38 @@ class PDFManager {
 
     // The PDF is not cached, extract it
     const cachePath = getCachePath(file)
-    const worker = PDFWorker.getWorker()
+    const data = new Uint8Array(await app.vault.readBinary(file))
 
-    return new Promise(async (resolve, reject) => {
-      try {
-        const res = await worker.run({ data: file.path, name: file.basename })
-        const text = (res.data.text as string)
-          // Replace \n with spaces
-          .replace(/\n/g, ' ')
-          // Trim multiple spaces
-          .replace(/ +/g, ' ')
-          .trim()
+    // Load the PDF.js library
+    const pdfjs = await loadPdfJs()
+    const loadingTask = pdfjs.getDocument({ data })
+    const pdf = await loadingTask.promise
+    const pagePromises = [];
+    // Get text from each page of the PDF
+    for (let j = 1; j <= pdf.numPages; j++) {
+      const page = pdf.getPage(j);
 
-        // Add it to the cache
-        await writeCache(cachePath.folder, cachePath.filename, text, file.path, '')
-        // Add a delay to prevent out-of-memory crash (hopefully garbage collection will run more often)
-        setTimeout(() => {
-          resolve(text)
-        }, 10000)
-      } catch (e) {
-        // In case of error (unreadable PDF or timeout) just add
-        // an empty string to the cache
-        await writeCache(cachePath.folder, cachePath.filename, '', file.path, '')
-        resolve('')
-      }
-    })
+      // @ts-ignore
+      pagePromises.push(page.then((page) => {
+        // @ts-ignore
+        const textContentPromise: Promise<{ items }> = page.getTextContent();
+        return textContentPromise.then((t) => {
+          const items = t.items;
+          let text = '';
+          for (let i = 0; i < items.length; i++) {
+            text += items[i].str + ' ';
+          }
+          return text;
+        });
+      }));
+    }
+
+    const texts = await Promise.all(pagePromises);
+
+    // Add it to the cache
+    const text = texts.join(' ');
+    await writeCache(cachePath.folder, cachePath.filename, text, file.path, '')
+    return text
   }
 }
